Add manual HRU refresh button with stern warning Fix HRU grid download name Don't show HRU Characteristics of Centralized Trash BMPs even though it's technically possible to calculate them

+++ Source/Neptune.Web/ScheduledJobs/LoadGeneratingUnitRefreshScheduledBackgrounDJob.cs
@@ -117,10 +117,14 @@ namespace Neptune.Web.ScheduledJobs
                 throw;
             }
 
-            File.Delete(outputLayerPath);
-            if (loadGeneratingUnitRefreshAreaID != null)
+            // clean up temp files if not running in a local environment
+            if (NeptuneWebConfiguration.NeptuneEnvironment.NeptuneEnvironmentType != NeptuneEnvironmentType.Local)
             {
-                File.Delete(clipLayerPath);
+                File.Delete(outputLayerPath);
+                if (loadGeneratingUnitRefreshAreaID != null)
+                {
+                    File.Delete(clipLayerPath);
+                }
             }
